- refactor rrc

+++ bfc/ITimer.hpp
@@ -9,8 +9,8 @@ namespace bfc
 
 struct ITimer
 {
-    virtual int schedule(std::chrono::nanoseconds pDiff, std::function<void()> pCb) = 0;
-    virtual void cancel(int) = 0;
+    virtual uint64_t schedule(std::chrono::nanoseconds pDiff, std::function<void()> pCb) = 0;
+    virtual void cancel(uint64_t) = 0;
 };
 
 } // namespace bfc
